requirements.txt
#Python==3.11.5
selenium== 4.15.0
beautifulsoup4==4.12.2
requests==2.31.0
pandas==2.1.2
Pillow==10.1.0
psycopg2==2.9.9
webdriver-manager==4.0.1
